backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.5
pydantic==2.8.2
google-generativeai==0.7.2
python-dotenv==1.0.0
Faker==26.0.0
httpx==0.27.2
agno==1.7.12
pyttsx3==2.90